Revert "[OWL-1644] fix error from toolkits --- CmdRunWithTimeout"
This reverts commit 21c0a0af26fd72b269bb461436fc967e678e7dc2.
