+++ modules/agent/http/plugin.go
@@ -4,13 +4,43 @@ import (
 	"fmt"
 	"github.com/Cepave/open-falcon-backend/modules/agent/g"
 	"github.com/Cepave/open-falcon-backend/modules/agent/plugins"
-	"github.com/Cepave/open-falcon-backend/modules/agent/session"
 	log "github.com/Sirupsen/logrus"
 	"github.com/toolkits/file"
 	"net/http"
 	"os/exec"
+	"syscall"
 	"time"
 )
+
+func sessionKill(cmd *exec.Cmd) error {
+	if err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL); err != nil {
+		log.Println("failed to kill processes of the same session: ", err)
+		return err
+	}
+	return nil
+}
+
+// cmdSessionRunWithTimeout runs cmd, and kills the children on timeout
+func cmdSessionRunWithTimeout(cmd *exec.Cmd, timeout time.Duration) (error, bool) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
+	log.Debugln("show cmd sysProcAttr.Setsid", cmd.SysProcAttr.Setsid)
+	if err := cmd.Start(); err != nil {
+		log.Errorln(cmd.Path, " start fails: ", err)
+		return err, false
+	}
+
+	done := make(chan error, 1)
+	go func() { done <- cmd.Wait() }()
+
+	select {
+	case <-time.After(timeout):
+		log.Printf("timeout, process:%s will be killed", cmd.Path)
+		err := sessionKill(cmd)
+		return err, true
+	case err := <-done:
+		return err, false
+	}
+}
 
 func configPluginRoutes() {
 	http.HandleFunc("/plugin/update", func(w http.ResponseWriter, r *http.Request) {
@@ -27,7 +57,7 @@ func configPluginRoutes() {
 			// git pull
 			cmd := exec.Command("git", "pull")
 			cmd.Dir = dir
-			err, isTimeout := session.CmdSessionRunWithTimeout(cmd, time.Duration(600)*time.Second)
+			err, isTimeout := cmdSessionRunWithTimeout(cmd, time.Duration(600)*time.Second)
 			if isTimeout {
 				// has be killed
 				if err == nil {
@@ -47,7 +77,7 @@ func configPluginRoutes() {
 			// git clone
 			cmd := exec.Command("git", "clone", g.Config().Plugin.Git, file.Basename(dir))
 			cmd.Dir = parentDir
-			err, isTimeout := session.CmdSessionRunWithTimeout(cmd, time.Duration(600)*time.Second)
+			err, isTimeout := cmdSessionRunWithTimeout(cmd, time.Duration(600)*time.Second)
 			if isTimeout {
 				// has be killed
 				if err == nil {
