+++ modules/agent/http/plugin_test.go
@@ -1,4 +1,4 @@
-package session
+package http
 
 import (
 	"github.com/toolkits/sys"
@@ -9,11 +9,18 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
+func TestDeleteAndCloneRepo(t *testing.T) {
+	out := DeleteAndCloneRepo("./plugin", "https://github.com/humorless/openfalcon-plugin.git")
+	t.Log("test deleteAndCloneRepo: ", out)
+	out = DeleteAndCloneRepo("/", "https://github.com/humorless/openfalcon-plugin.git")
+	t.Log("test deleteAndCloneRepo: ", out)
+}
+
 func TestRunCmdFamilyWithTimeout(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
 	t1 := time.Now()
 	cmd := exec.Command("/bin/sh", "-c", "watch date > date.txt")
-	err, isTimeout := CmdSessionRunWithTimeout(cmd, time.Duration(3)*time.Second)
+	err, isTimeout := cmdSessionRunWithTimeout(cmd, time.Duration(3)*time.Second)
 	t2 := time.Now()
 	t.Log("Time spent should less than 4 second: ", t2.Sub(t1))
 	t.Log("error message is:", err)
@@ -26,7 +33,7 @@ func TestRunCmdFamilyWithTimeout(t *testing.T) {
 	t.Log("There should not be any 'watch date' existing.")
 
 	cmd2 := exec.Command("sleep", "1")
-	err2, isTimeout2 := CmdSessionRunWithTimeout(cmd2, time.Duration(3)*time.Second)
+	err2, isTimeout2 := cmdSessionRunWithTimeout(cmd2, time.Duration(3)*time.Second)
 	t.Log("error message is:", err2)
 	t.Log("Timeout happens: ", isTimeout2)
 	if isTimeout2 != false {
