+++ modules/agent/plugins/scheduler.go
@@ -6,9 +6,9 @@ import (
 	"encoding/json"
 	"github.com/Cepave/open-falcon-backend/common/model"
 	"github.com/Cepave/open-falcon-backend/modules/agent/g"
-	"github.com/Cepave/open-falcon-backend/modules/agent/session"
 	log "github.com/Sirupsen/logrus"
 	"github.com/toolkits/file"
+	"github.com/toolkits/sys"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -133,7 +133,7 @@ func PluginRun(plugin *Plugin) {
 		log.Errorln(fpath, " start fails: ", err)
 	}
 
-	err, isTimeout := session.CmdSessionRunWithTimeout(cmd, time.Duration(timeout)*time.Millisecond)
+	err, isTimeout := sys.CmdRunWithTimeout(cmd, time.Duration(timeout)*time.Millisecond)
 
 	errStr := stderr.String()
 	if errStr != "" {
